前段控制任务开始结束时间

+++ shared/src/main/scala/org/ncgw/noron/shared/TimeLineProtocol.scala
@@ -7,7 +7,7 @@ package org.ncgw.noron.shared
   */
 object TimeLineProtocol {
   case class TimeLineItem(
-                         taskId: Int,
+                         taskId: Long,
                          startTime: Long,
                          endTime: Long,
                          content: String,
